Move json-server files to appropriate folder Rename dashboard module files, export and fix import Rename teravoz_client folder to client
