--- frontend/pages/components/dashboard/index.js
@@ -1,3 +1,0 @@
-import Dashboard from "./dashboard";
-
-export default Dashboard;
