+++ package.json
@@ -8,9 +8,9 @@
     "dashboard:dev": "node_modules/next/dist/bin/next frontend/ -p 5000",
     "dashboard:build": "node_modules/next/dist/bin/next build frontend/",
     "dashboard:start": "node_modules/next/dist/bin/next start frontend/ -H 0.0.0.0 -p 5000",
-    "api:teravoz": "nodemon --exec 'babel-node teravoz/server.js' --ignore db.json --ignore frontend/",
-    "client": "nodemon --exec 'babel-node server.js' --ignore db.json --ignore frontend/",
-    "jsonServer": "json-server --host 0.0.0.0 --watch db.json --routes json-server-routes.js",
+    "api:teravoz": "nodemon --exec 'babel-node teravoz/server.js' --ignore jsonServer/db.json --ignore frontend/",
+    "client": "nodemon --exec 'babel-node client/server.js' --ignore jsonServer/db.json --ignore frontend/",
+    "jsonServer": "json-server --host 0.0.0.0 --watch jsonServer/db.json --routes jsonServer/json-server-routes.js",
     "dev": "npm-run-all --parallel jsonServer api:teravoz client dashboard:dev",
     "production": "yarn dashboard:build && npm-run-all --parallel jsonServer api dashboard:start",
     "test:watch": "jest --watchAll --verbose",
